+++ askama_derive/Cargo.toml
@@ -18,4 +18,6 @@ rocket = ["askama_shared/rocket"]
 
 [dependencies]
 askama_shared = { version = "0.6.0", path = "../askama_shared" }
+nom = "3"
+quote = "0.5"
 syn = "0.13"
