+++ askama_shared/Cargo.toml
@@ -16,8 +16,5 @@ rocket = []
 
 [dependencies]
 error-chain = "0.11"
-nom = "3"
-quote = "0.5"
 serde = { version = "1.0", optional = true }
 serde_json = { version = "1.0", optional = true }
-syn = "0.13"
